v0.1.0

+++ src/applications/stores/campaignStore.ts
@@ -6,6 +6,7 @@ import { watch, ref, computed } from 'vue';
 
 // local imports
 import { useCampaignDirectoryStore, useMainStore, useNavigationStore } from '@/applications/stores';
+import { confirmDialog } from '@/dialogs';
 
 // types
 import { PCDetails, FieldData, CampaignLoreDetails} from '@/types';
@@ -76,6 +77,10 @@ export const useCampaignStore = defineStore('campaign', () => {
     if (!pc) 
       throw new Error('Bad session in campaignDirectoryStore.deletePC()');
 
+    // confirm
+    if (!(await confirmDialog('Delete PC?', 'Are you sure you want to delete this PC?')))
+      return;
+
     await pc.delete();
 
     // update tabs/bookmarks
@@ -86,14 +91,17 @@ export const useCampaignStore = defineStore('campaign', () => {
   };
   
     /**
-   * Adds a lore to the session.
+   * Adds a lore to the campaign.
+   * @param description The description for the lore entry
+   * @returns The UUID of the created lore entry
    */
-    const addLore = async (description = ''): Promise<void> => {
+    const addLore = async (description = ''): Promise<string | null> => {
       if (!currentCampaign.value)
-        throw new Error('Invalid session in campaignStore.addLore()');
+        throw new Error('Invalid campaign in campaignStore.addLore()');
   
-      await currentCampaign.value.addLore(description);
+      const loreUuid = await currentCampaign.value.addLore(description);
       await _refreshLoreRows();
+      return loreUuid;
     }
   
     /**
@@ -123,12 +131,16 @@ export const useCampaignStore = defineStore('campaign', () => {
   
     /**
      * Deletes a lore from the session
-     * @param uuid the UUID of the l0ore
+     * @param uuid the UUID of the lore
      */
     const deleteLore = async (uuid: string): Promise<void> => {
       if (!currentCampaign.value)
         throw new Error('Invalid session in campaignStore.deleteLore()');
   
+      // confirm
+      if (!(await confirmDialog('Delete Lore?', 'Are you sure you want to delete this lore?')))
+        return;
+
       await currentCampaign.value.deleteLore(uuid);
       await _refreshLoreRows();
     }
@@ -303,7 +315,7 @@ export const useCampaignStore = defineStore('campaign', () => {
         let entry: JournalEntryPage | null = null;
 
         if (lore.journalEntryPageId)
-          entry = await fromUuid(lore.journalEntryPageId) as JournalEntryPage;
+          entry = await fromUuid(lore.journalEntryPageId) as JournalEntryPage | null;
   
         retval.push({
           uuid: lore.uuid,
@@ -324,7 +336,7 @@ export const useCampaignStore = defineStore('campaign', () => {
       let entry: JournalEntryPage | null = null;
 
       if (lore.journalEntryPageId)
-        entry = await fromUuid(lore.journalEntryPageId) as JournalEntryPage;
+        entry = await fromUuid(lore.journalEntryPageId) as JournalEntryPage | null;
 
       retval.push({
         uuid: lore.uuid,
